chore: regenerate with oagen-emitter dotnet fixes
Cumulative regen from the emitter-side work (P0.1, P0.3, P0.4, P1.8–10,
P2.11–14, P3.15–16) plus the IR-level enum synthesis (P0.2). Notable
user-visible changes:

- Required enum fields skip serializing the Unknown sentinel so unset
  required enums produce a clean missing-field API error instead of
  `"unknown"` on the wire.
- Service methods collapse to one-line GetAsync/PostAsync/… calls on
  the Service base; hundreds of inlined WorkOSRequest blocks removed.
- Duplicate enum classes collapse to canonical names; parent classes
  that only differed by aliased children dedupe further.
- Literal / single-value enum fields ("object": "event", …) emit as
  string with an internal setter and const initializer.
- Dictionary<string, object> fields get typed Get<Field>Attribute<T>
  accessors that work under Newtonsoft and STJ.
- oneOf-of-string-consts (e.g. UserIdentitiesGetItem.provider) now
  materialize as a real enum (UserIdentitiesGetItemProvider) instead
  of collapsing to `object`.
- Generated tests seed required string body/query fields and assert
  against the exact fixture values, catching snake_case regressions.
- URL-builder ops (GetAuthorizationUrl, GetLogoutUrl) get sync URL-
  structure tests instead of awaited HttpMock calls.
- XML doc summaries preserve acronyms (SSO, API, …) and pick the
  right indefinite article.

+++ src/WorkOS.net/Entities/UpdateAuditLogsRetention.cs
@@ -5,13 +5,13 @@ namespace WorkOS
     using Newtonsoft.Json;
     using STJS = System.Text.Json.Serialization;
 
-    /// <summary>Represents a update audit logs retention.</summary>
+    /// <summary>Represents an update audit logs retention.</summary>
     public class UpdateAuditLogsRetention
     {
 
-        /// <summary>The number of days Audit Log events will be retained. Valid values are `30` and `365`.</summary>
+        /// <summary>The number of days to retain audit logs. Valid values are 30 or 365.</summary>
         [JsonProperty("retention_period_in_days")]
         [STJS.JsonPropertyName("retention_period_in_days")]
-        public long RetentionPeriodInDays { get; set; }
+        public double RetentionPeriodInDays { get; set; }
     }
 }
